requirements.txt
ffmpeg-python
exifread
pyinstaller